@@ -1,107 +1,79 @@
 name: bbnaf
 description: Bloodbowl tournament management App.
-
-# The following line prevents the package from being accidentally published to
-# pub.dev using `pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 1.1.0+1
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
+  adaptive_dialog: ^2.0.0
+  authentication_repository:
+    path: packages/authentication_repository
+  bloc: ^8.1.0
+  bloc_concurrency: ^0.2.3
+  cache:
+    path: packages/cache
+  checkbox_formfield: ^0.2.0
+  cloud_firestore: ^4.3.2
+  cupertino_icons: ^1.0.0
+  data_tables: ^1.0.3
+  data_table_2: ^2.5.10
+  english_words: ^4.0.0
+  enum_to_string: ^2.0.1
+  equatable: ^2.0.3
+  excel: ^4.0.2
+  file_picker: ^6.1.1
+  firebase_auth: ^4.2.6
+  firebase_core: ^2.5.0
+  firebase_messaging: ^14.2.4
+  firebase_storage: ^11.0.11
+  firebase_storage_web: ^3.3.20
+  flow_builder: ^0.0.9
   flutter:
     sdk: flutter
-
-  uuid: ^4.0.0
-  cupertino_icons: ^1.0.0
-  english_words: ^4.0.0
-  json_annotation: ^4.0.1
-  intl: ^0.17.0
-  equatable: ^2.0.3
-  flutter_bloc: ^7.2.0
-  firebase_core: ^2.5.0
-  firebase_auth: ^4.2.6
-  cloud_firestore: ^4.3.2
-  firebase_storage: ^11.0.11
-  universal_html: ^2.0.8
-  firebase_storage_web: ^3.3.20
-  firebase_messaging: ^14.2.4
-  xml: ^5.3.0
-  flutter_login: ^4.1.1
-  adaptive_dialog: ^1.8.0+1
-  flutter_html: ^2.2.1
+  flutter_bloc: ^8.1.3
+  flutter_login: ^5.0.0 # could be removed?
+  flutter_native_splash: ^2.3.9
   flutter_rating_bar: ^4.0.1
+  flutter_widget_from_html: ^0.14.11
+  fluttertoast: ^8.1.4
+  font_awesome_flutter: ^10.1.0
+  form_inputs:
+    path: packages/form_inputs
+  formz: ^0.6.0
+  google_fonts: ^4.0.0
+  grouped_list: ^5.1.2
+  intl: ^0.19.0
+  json_annotation: ^4.0.1
+  json_serializable: ^6.7.1
+  meta: ^1.7.0
+  responsive_framework: ^1.1.1
+  string_similarity: ^2.0.0
   uni_links: ^0.5.1
-  responsive_framework: ^0.2.0
+  uuid: ^4.0.0
+  xml: ^6.5.0
+
+
+  # universal_html: ^2.0.8
+  # flutter_html: ^2.2.1
   # data_table_2: ^2.2.1
   # syncfusion_flutter_datepicker: ^19.2.44
-  json_serializable: ^4.1.3
-  pluto_grid: ^2.10.3
-  data_tables: ^1.0.3
-  grouped_list: ^4.0.0
-  flutter_native_splash: ^1.2.0
-  fluttertoast: ^8.1.2
-  file_picker: ^5.2.5
-  enum_to_string: ^2.0.1
+  # pluto_grid: ^7.0.2
+  
 
 dev_dependencies:
+  bloc_test: ^9.0.0
+  build_runner: ^2.0.6  
   flutter_test:
     sdk: flutter
+  mocktail: ^1.0.0
 
-  build_runner: ^2.0.6  
-
-
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
 
 # The following section is specific to Flutter.
 flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
-
-  # To add assets to your application, add an assets section, like this:
   assets:
     - assets/images/background/
     - assets/images/logos/
     - assets/images/teams/
-
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
