@@ -1,4 +1,4 @@
-import 'package:bbnaf/models/tournament/tournament.dart';
+import 'package:bbnaf/tournament_repository/src/models/models.dart';
 
 // Identifies if the object is of type squad or coach
 enum OrgType {
