@@ -1,25 +1,25 @@
-import 'package:bbnaf/repos/auth/auth_user.dart';
+// import 'package:bbnaf/repos/auth/auth_user.dart';
 
-abstract class AuthRepository {
-  Future<AuthUser> signInWithCredentials(String email, String password);
+// abstract class AuthRepository {
+//   Future<AuthUser> signInWithCredentials(String email, String password);
 
-  Future<AuthUser> signUp(
-      {required String nafName,
-      required String email,
-      required String password});
+//   Future<AuthUser> signUp(
+//       {required String nafName,
+//       required String email,
+//       required String password});
 
-  void signIn(String nafName);
+//   void signIn(String nafName);
 
-  Future<void> signOut();
+//   Future<void> signOut();
 
-  Future<bool> isSignedIn();
+//   Future<bool> isSignedIn();
 
-  // Returns null if user not signed in
-  String? getUserDisplayName();
+//   // Returns null if user not signed in
+//   String? getUserDisplayName();
 
-  // Returns null if user not signed in
-  String? getUserEmail();
+//   // Returns null if user not signed in
+//   String? getUserEmail();
 
-  // Returns current user (internals are null if not signed in)
-  AuthUser getAuthUser();
-}
+//   // Returns current user (internals are null if not signed in)
+//   AuthUser getAuthUser();
+// }
