@@ -1,27 +1,27 @@
-import 'package:firebase_auth/firebase_auth.dart';
+// import 'package:firebase_auth/firebase_auth.dart';
 
-class AuthUser {
-  User? user;
-  String? _nafName;
-  String? error;
+// class AuthUser {
+//   User? user;
+//   String? _nafName;
+//   String? error;
 
-  AuthUser({this.user, this.error}) {
-    this._nafName = this.user?.displayName;
-  }
+//   AuthUser({this.user, this.error}) {
+//     this._nafName = this.user?.displayName;
+//   }
 
-  AuthUser.nafNameOnly(String nafName) {
-    this._nafName = nafName.trim().toLowerCase();
-  }
+//   AuthUser.nafNameOnly(String nafName) {
+//     this._nafName = nafName.trim().toLowerCase();
+//   }
 
-  String getNafName() {
-    if (_nafName == null) {
-      _nafName = user?.displayName;
-    }
+//   String getNafName() {
+//     if (_nafName == null) {
+//       _nafName = user?.displayName;
+//     }
 
-    return _nafName != null ? _nafName!.trim() : "";
-  }
+//     return _nafName != null ? _nafName!.trim() : "";
+//   }
 
-  String getEmail() {
-    return user != null && user!.email != null ? user!.email as String : "";
-  }
-}
+//   String getEmail() {
+//     return user != null && user!.email != null ? user!.email as String : "";
+//   }
+// }
